requirements.txt
jina==3.15.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.4.3
pyyaml==6.0.1
langchain==0.0.335